feat(app): add move information
